Fix clippy complain about sleep mod having the same name as its parent
